+++ src/note.cpp
@@ -115,17 +115,6 @@ void Note::Reset()
     m_MIDIPitch = -1;
 }
 
-bool Note::HasToBeAligned() const
-{
-    if (!this->IsInLigature()) return true;
-    Note *note = const_cast<Note *>(this);
-    Ligature *ligature = dynamic_cast<Ligature *>(note->GetFirstAncestor(LIGATURE));
-    assert(ligature);
-    Note *firstNote = dynamic_cast<Note *>(ligature->GetList(ligature)->front());
-    Note *lastNote = dynamic_cast<Note *>(ligature->GetList(ligature)->back());
-    return ((note == firstNote) || (note == lastNote));
-}
-
 void Note::AddChild(Object *child)
 {
     // additional verification for accid and artic - this will no be raised with editorial markup, though
@@ -993,29 +982,37 @@ int Note::GenerateMIDI(FunctorParams *functorParams)
         return FUNCTOR_SIBLINGS;
     }
 
-    // Create midi this
-    int midiBase = 0;
-    data_PITCHNAME pname = note->GetPname();
-    switch (pname) {
-        case PITCHNAME_c: midiBase = 0; break;
-        case PITCHNAME_d: midiBase = 2; break;
-        case PITCHNAME_e: midiBase = 4; break;
-        case PITCHNAME_f: midiBase = 5; break;
-        case PITCHNAME_g: midiBase = 7; break;
-        case PITCHNAME_a: midiBase = 9; break;
-        case PITCHNAME_b: midiBase = 11; break;
-        case PITCHNAME_NONE: break;
+    int pitch = 0;
+    if (note->HasPnum()) {
+        pitch = note->GetPnum();
     }
-    // Check for accidentals
-    midiBase += note->GetChromaticAlteration();
+    else {
+        // calc pitch
+        int midiBase = 0;
+        data_PITCHNAME pname = note->GetPname();
+        if (note->HasPnameGes()) pname = note->GetPnameGes();
+        switch (pname) {
+            case PITCHNAME_c: midiBase = 0; break;
+            case PITCHNAME_d: midiBase = 2; break;
+            case PITCHNAME_e: midiBase = 4; break;
+            case PITCHNAME_f: midiBase = 5; break;
+            case PITCHNAME_g: midiBase = 7; break;
+            case PITCHNAME_a: midiBase = 9; break;
+            case PITCHNAME_b: midiBase = 11; break;
+            case PITCHNAME_NONE: break;
+        }
+        int oct = note->GetOct();
+        if (note->HasOctGes()) oct = note->GetOctGes();
 
-    // Adjustment for transposition intruments
-    midiBase += params->m_transSemi;
+        // Check for accidentals
+        midiBase += note->GetChromaticAlteration();
 
-    int oct = note->GetOct();
-    if (note->HasOctGes()) oct = note->GetOctGes();
+        // Adjustment for transposition intruments
+        midiBase += params->m_transSemi;
 
-    int pitch = midiBase + (oct + 1) * 12;
+        pitch = midiBase + (oct + 1) * 12;
+    }
+
     // We do store the MIDIPitch in the note even with a sameas
     this->SetMIDIPitch(pitch);
     int channel = params->m_midiChannel;