+++ src/toolkit.cpp
@@ -1257,6 +1257,8 @@ std::string Toolkit::RenderToSVG(int pageNo, bool xml_declaration)
         svg.SetSvgViewBox(true);
     }
 
+    svg.SetHtml5(m_options->m_svgHtml5.GetValue());
+
     // render the page
     RenderToDeviceContext(pageNo, &svg);
 
@@ -1486,8 +1488,10 @@ std::string Toolkit::GetMIDIValuesForElement(const std::string &xmlId)
         assert(note);
         int timeofElement = this->GetTimeForElement(xmlId);
         int pitchofElement = note->GetMIDIPitch();
+        int durationofElement = note->GetScoreTimeDuration();
         o << "time" << timeofElement;
         o << "pitch" << pitchofElement;
+        o << "duration" << durationofElement;
     }
     return o.json();
 }