+++ src/view_mensural.cpp
@@ -85,7 +85,9 @@ void View::DrawMensuralNote(DeviceContext *dc, LayerElement *element, Layer *lay
     // Semibrevis and shorter
     else {
         wchar_t code = note->GetMensuralSmuflNoteHead();
+        dc->StartCustomGraphic("notehead");
         DrawSmuflCode(dc, xNote, yNote, code, staff->m_drawingStaffSize, false);
+        dc->EndCustomGraphic();
         // For semibrevis with stem in black notation, encoded with an explicit stem direction
         if (((drawingDur > DUR_1) || (note->GetStemDir() != STEMDIRECTION_NONE))
             && note->GetStemVisible() != BOOLEAN_false) {
@@ -334,12 +336,18 @@ void View::DrawMaximaToBrevis(DeviceContext *dc, int y, LayerElement *element, L
         if (note->GetStemDir() != STEMDIRECTION_NONE) {
             up = (note->GetStemDir() == STEMDIRECTION_up);
         }
+        else if (staff->m_drawingNotationType == NOTATIONTYPE_NONE
+            || staff->m_drawingNotationType == NOTATIONTYPE_cmn) {
+            up = (note->GetDrawingStemDir() == STEMDIRECTION_up);
+        }
         shape = (up) ? LIGATURE_STEM_RIGHT_UP : LIGATURE_STEM_RIGHT_DOWN;
     }
 
     Point topLeft, bottomRight;
     int sides[4];
     this->CalcBrevisPoints(note, staff, &topLeft, &bottomRight, sides, shape, isMensuralBlack);
+
+    dc->StartCustomGraphic("notehead");
 
     if (!fillNotehead) {
         // double the bases of rectangles
@@ -353,7 +361,19 @@ void View::DrawMaximaToBrevis(DeviceContext *dc, int y, LayerElement *element, L
 
     // serifs and / or stem
     DrawFilledRectangle(dc, topLeft.x, sides[0], topLeft.x + stemWidth, sides[1]);
-    DrawFilledRectangle(dc, bottomRight.x - stemWidth, sides[2], bottomRight.x, sides[3]);
+
+    if (note->GetActualDur() != DUR_BR) {
+        // Right side is a stem - end the notehead first
+        dc->EndCustomGraphic();
+        dc->StartCustomGraphic("stem");
+        DrawFilledRectangle(dc, bottomRight.x - stemWidth, sides[2], bottomRight.x, sides[3]);
+        dc->EndCustomGraphic();
+    }
+    else {
+        // Right side is a serif
+        DrawFilledRectangle(dc, bottomRight.x - stemWidth, sides[2], bottomRight.x, sides[3]);
+        dc->EndCustomGraphic();
+    }
 
     return;
 }