+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## [unreleased]
+* Support for `@pname.ges` and `@pnum` in MIDI output (@rettinghaus)
+* Support for `@xml:space` in MusicXML import (@rettinghaus)
+* Support for `staffDef@lines.visible` (@rettinghaus)
+* Support for gliss (@rettinghaus)
+* Improved bTrem handling (@rettinghaus)
 * Improved trill support (@rettinghaus)
 * Option for using encoded line breaks, but automatic page breaks (--breaks=line) (@earboxer)
 * Support for `trill@extender` (@rettinghaus)